+++ contracts/wallet/src/message.rs
@@ -1,5 +1,6 @@
 use core::time::Duration;
 
+use defuse_nep641::OffchainMessage;
 use defuse_time::Timestamp;
 use near_account_id::AccountId;
 
@@ -207,6 +208,71 @@ impl RequestMessage {
             return None;
         }
         self.deadline().duration_since(now).ok()
+    }
+}
+
+/// NEP-641 authorization for [`Wallet`](crate::contract::Wallet) contract.
+#[cfg_attr(
+    feature = "serde",
+    derive(::serde::Serialize, ::serde::Deserialize),
+    cfg_attr(feature = "schemars-v0_8", derive(::schemars::JsonSchema)),
+    serde(rename_all = "snake_case")
+)]
+#[cfg_attr(feature = "arbitrary", derive(::arbitrary::Arbitrary))]
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub enum WalletAuthorization {
+    Signature {
+        msg: OffchainMessage,
+        proof: String,
+    },
+    Extension {
+        account_id: AccountId,
+        authorization: String,
+        payload: String,
+    },
+}
+
+impl WalletAuthorization {
+    /// Get the authorized payload
+    #[inline]
+    pub const fn payload(&self) -> &str {
+        match self {
+            Self::Signature {
+                msg: OffchainMessage { payload, .. },
+                ..
+            } => payload.as_str(),
+            Self::Extension { payload, .. } => payload.as_str(),
+        }
+    }
+
+    /// Convert into the authorized payload
+    #[inline]
+    pub fn into_payload(self) -> String {
+        match self {
+            Self::Signature {
+                msg: OffchainMessage { payload, .. },
+                ..
+            } => payload,
+            Self::Extension { payload, .. } => payload,
+        }
+    }
+
+    /// Convert to the authorization blob
+    #[cfg(feature = "json")]
+    #[inline]
+    pub fn to_authorization(&self) -> String {
+        serde_json::to_string(self).expect("JSON: failed to serialize")
+    }
+
+    /// Wrap as extension with given ID
+    #[cfg(feature = "json")]
+    #[inline]
+    pub fn as_extension_of(self, account_id: impl Into<AccountId>) -> Self {
+        Self::Extension {
+            account_id: account_id.into(),
+            authorization: self.to_authorization(),
+            payload: self.into_payload(),
+        }
     }
 }
